+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Minor: Added a description for the logging option under moderation tab. (#6514)
 - Minor: Added a menu action to sort tabs alphabetically. (#6551)
 - Minor: Fixed "edit hotkey" dialog opening like a normal window. (#6540)
+- Minor: Added an API to get the current Twitch account in plugins. (#6554)
 - Bugfix: Expose the "Extra extension IDs" setting on non-Windows systems too. (#6509)
 - Bugfix: Fixed some commands and filters not working as expected in seach popups. (#6539)
 - Bugfix: Fixed settings occasionally not opening when clicking on "Manage Accounts" in the account switcher. (#6543)